Adding support to transfer amount in decimals

+++ money_transfer/src/main/java/com/revolut/money_transfer/request/MoneyTransferRequest.java
@@ -8,13 +8,13 @@ public class MoneyTransferRequest {
 	
 	private Integer toAccountId;
 	
-	private Integer amount;
+	private Double amount;
 	
 	public MoneyTransferRequest() {
 		super();
 	}
 	
-	public MoneyTransferRequest(Integer fromAccountId, Integer toAccountId, Integer amount) {
+	public MoneyTransferRequest(Integer fromAccountId, Integer toAccountId, Double amount) {
 		this.fromAccountId = fromAccountId;
 		this.toAccountId = toAccountId;
 		this.amount = amount;
@@ -36,11 +36,11 @@ public class MoneyTransferRequest {
 		this.toAccountId = toAccountId;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
